Modularize CloudSync adapters
Separate credential exchange, native token configuration, and session eviction persistence from the lifecycle state machine without changing transition policy.
